backend/requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
uvicorn==0.24.0.post1
python-dotenv==1.0.0
pydantic-settings==2.0.3
pyjwt[crypto]==2.8.0
sqlalchemy==2.0.23
alembic~=1.12.1
pydantic~=2.4.2